Add launcher, launcher-client, and launcher-protocol crates
This changes the process model for running the supervisor and how
the supervisor spawns and supervises running processes. It adds a
new binary (`hab-launch`) which, when started, spawns the latest
Habitat supervisor which is already installed and passes the
arguments it recieved to the Supervisor itself. If the Supervisor
crashes or shuts itself down then the Launcher will restart it.

The Supervisor no longer spawns child proccesses itself. It will
now delegate that task to the Launcher by communicating over a
bi-directional named pipe using protobufs which is established
at startup.

These changes allow the Supervisor to be shut down for an upgrade
without shutting down any of the running processes. The Launcher
itself is a very small Rust program which is intentionally stupid.
If you think you should add something to the Launcher you should
very much question your intents and talk it over with as many
Habitat core maintainers as possible :).

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>